requirements.txt
absl-py==2.1.0
adlfs==2024.7.0
aiobotocore==2.13.3
aiohappyeyeballs==2.4.0
aiohttp==3.10.11
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
azure-datalake-store==0.0.53
azure-identity==1.17.1
azure-storage-blob==12.22.0
backports.tarfile==1.2.0
bitsandbytes==0.43.3
botocore==1.34.162
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
croniter==3.0.3
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.5.9
decorator==5.1.1
diskcache==5.6.3
dm-tree==0.1.8
docker==7.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
filelock==3.15.4
flyteidl==1.13.2
flytekit==1.13.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gcsfs==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.66.0
grpcio-status==1.66.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.4.0
importlib_resources==6.4.4
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.32.1
jsonlines==4.0.0
jsonpickle==3.2.2
keyring==25.3.0
kiwisolver==1.4.5
lightning==2.4.0
lightning-utilities==0.11.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.13.0
mashumaro==3.13.1
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.4.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==2.1.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.41.0
nvidia-nccl-cu12==2.20.5
nvidia-nvimgcodec-cu12==0.3.0.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
piq==0.8.0
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.3.0
proto-plus==1.24.0
protobuf==5.27.3
protoc-gen-openapiv2==0.0.1
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytimeparse==1.1.8
pytorch-lightning==2.3.0
pytorchcv==0.0.73
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rich-click==1.8.3
rsa==4.9
s3fs==2024.6.1
seaborn==0.13.2
SecretStorage==3.3.3
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==75.8.2
six==1.16.0
smmap==5.0.1
statsd==4.0.1
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
torch==2.4.0
torch-tb-profiler==0.4.3
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
triton==3.0.0
typeshed_client==2.7.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.18.1
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.16.0
yarl==1.18.3
zipp==3.20.0
